\name{author_poiscon}
\alias{author_poiscon}
\title{Author poiscon}
\usage{
author_poiscon(deck_dir = "output/deck",
  jekyll_dir = "poissonconsulting.github.io",
  jekyll_path = "~/Documents/code")
}
\arguments{
  \item{dir}{string of github repository name.}
}
\description{
Creates slidify deck
}

