\name{dataset.ggplot}
\alias{dataset.ggplot}
\title{Get dataset}
\usage{
\method{dataset}{ggplot}(object, ...)
}
\arguments{
  \item{object}{a \code{ggplot} object.}

  \item{...}{further arguments passed to or from other
  methods.}
}
\value{
The dataset.
}
\description{
Gets the dataset from a \code{ggplot} object.
}

