\name{deck_to_web}
\alias{deck_to_web}
\title{Deck to web repository}
\usage{
deck_to_web(release_date = NULL, dir = "deck",
  web_dir = "poissonconsulting.github.io")
}
\arguments{
  \item{dir}{a string of the slide deck directory}

  \item{release_date}{a Date scalar of the date the deck
  was finalized or NULL}

  \item{web_dir}{a string of the web repository}
}
\value{
deck to web
}
\description{
Copies slide deck to jekyll site
}

