\name{delete_output}
\alias{delete_output}
\title{Delete output folder}
\usage{
delete_output(check = TRUE)
}
\arguments{
  \item{check}{a logical scalar indicating whether to check
  with user before deleting}
}
\value{
An invisible logical scalar indicating if successful.
}
\description{
Deletes output folder for current project.
}

