\name{extract_figures}
\alias{extract_figures}
\title{Extract figures}
\usage{
extract_figures()
}
\value{
An invisible logical scalar indicating whether or not
sucessful.
}
\description{
Extracts all the png files from output/plots/figures and
saves to output/report/figures/project-name ready for
linking to the report and posting on the jekyll site in the
figures directory.
}

