\name{figures_report}
\alias{figures_report}
\title{Extract figures}
\usage{
figures_report()
}
\value{
An invisible logical scalar indicating whether or not
sucessful.
}
\description{
Extracts all the png files from output/plots/figures and
saves to output/report/project-name/figures ready for
linking to the report and posting on the jekyll site in the
figures directory.
}

