\name{gwindow}
\alias{gwindow}
\title{Create new plotting window}
\usage{
  gwindow(dwidth = 1, dheight = 1, slide = FALSE)
}
\arguments{
  \item{dwidth}{the proportion of the page width}

  \item{dheight}{the proprotion of the page height}

  \item{slide}{whether or not powerpoint slide (TRUE)
  versus plot}
}
\description{
  Creates a new window for plotting
}

