\name{h}
\alias{h}
\title{Head}
\usage{
  h(x, ...)
}
\arguments{
  \item{x}{an object to get the head}

  \item{...}{further arguments passed to or from other
  methods.}
}
\value{
  heads x.
}
\description{
  Short hand for head
}
\seealso{
  \code{\link{p}}, \code{\link{s}}, \code{\link{ht}}.
}

