\name{ht}
\alias{ht}
\title{Head-tail}
\usage{
  ht(x, ...)
}
\arguments{
  \item{x}{an object to get the head and tail}

  \item{...}{further arguments passed to or from other
  methods.}
}
\value{
  Head and tail of x.
}
\description{
  Short hand for head and tail
}
\seealso{
  \code{\link{p}}, \code{\link{s}}, \code{\link{h}}.
}

