\name{images_web}
\alias{images_web}
\title{Images to web}
\usage{
images_web(dir = "poissonconsulting.github.io")
}
\arguments{
  \item{dir}{jekyll repository}
}
\value{
An invisible logical scalar indicating whether or not
sucessful.
}
\description{
Transfers all the files output/report/project-name/images
to the project folder in the images directory on the jekyll
site
}

