\name{increment_version_number}
\alias{increment_version_number}
\title{Increment version number}
\usage{
increment_version_number()
}
\value{
logical scalar indicating whether successful
}
\description{
Increments description files version number and date
}
\references{
Modified from
http://rmflight.github.io/posts/2014/02/incrementalVersionCommitHooks.html
}

