\name{is_informative_prediction}
\alias{is_informative_prediction}
\title{Is informative prediction}
\usage{
is_informative_prediction(x)
}
\arguments{
  \item{x}{data.frame of predictions to test informative}
}
\value{
logical scalar indicating whether informative
}
\description{
Tests whether a prediction data.frame is informative i.e.,
has variation in estimate and/or lower/upper.
}

