\name{knit_figures}
\alias{knit_figures}
\title{Knit figures}
\usage{
knit_figures(replacement = NULL)
}
\arguments{
  \item{replacement}{an optional named character vector
  which specifies replacement names for directories as well
  as the order to print the estimates.}
}
\value{
Embeds figures in markdown format.
}
\description{
Takes all plots in figures folder and embeds in markdown
format using html.
}

