\name{knit_report}
\alias{knit_report}
\title{Knit Report}
\usage{
knit_report(report = "report.Rmd")
}
\arguments{
  \item{report}{a character scalar indicating the RMarkdown
  file to knit.}
}
\value{
An invisible logical scalar indicating if successful.
}
\description{
Knits report file. An inhouse function.
}

