\name{list_assumptions}
\alias{list_assumptions}
\title{List assumptions}
\usage{
list_assumptions(model)
}
\arguments{
  \item{model}{character scalar of analysis part of model
  name}
}
\value{
logical scalar indicating whether successful
}
\description{
Takes assumptions character vector in file models-'model'.r
and cats as unordered list in markdown
}

