% Generated by roxygen2 (4.0.2): do not edit by hand
\name{load_analysis}
\alias{load_analysis}
\title{Load analysis}
\usage{
load_analysis(name = "analysis")
}
\arguments{
\item{name}{the name of the file (without extension).}
}
\value{
an analysis object
}
\description{
Reads in .rdata file from current analysis folder.
}

