\name{load_pkg}
\alias{load_pkg}
\title{Loads package from hard drive}
\usage{
load_pkg(pkg = "poiscon", code_dir = options()$code_dir)
}
\arguments{
  \item{pkg}{string of repository directory name}

  \item{code_dir}{string of path to repository directory}
}
\value{
Loads package
}
\description{
Loads package from hard drive
}

