\name{load_plot}
\alias{load_plot}
\title{Load plot}
\usage{
load_plot(name = "plot", type = "figures")
}
\arguments{
  \item{name}{character scalar of the name of the plot
  (without extension).}

  \item{type}{the plot type i.e. figures (default) or
  analysis, data etc.}
}
\value{
a data.frame object
}
\description{
Reads in ggplot object
}

