% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/load.r
\name{load_rdata}
\alias{load_rdata}
\title{Load rdata}
\usage{
load_rdata(name = NULL)
}
\arguments{
\item{name}{the name of the file (without extension) to load. If
unspecified then loads all R data files in current rdata folder
into the global environment.}
}
\value{
the object if name is specified otherwise loads
all objects in current rdata folder into global workspace
}
\description{
Reads in R data file from current rdata folder.
}

