\name{load_table}
\alias{load_table}
\title{Load table}
\usage{
  load_table(name = "table", type = "results")
}
\arguments{
  \item{name}{the name of the table (without extension).}

  \item{type}{the table type i.e. results (default) or
  analysis etc.}
}
\value{
  a data.frame object
}
\description{
  Reads in table.csv file from current table folder.
}

