% Generated by roxygen2 (4.0.2): do not edit by hand
\name{load_value}
\alias{load_value}
\title{Load table}
\usage{
load_value(rowname, colname = "estimate", folders = get_folders(),
  name = "estimates", type = "results", digits = 6)
}
\arguments{
\item{rowname}{string of name of row containing value}

\item{colname}{string of name of column containing value}

\item{folders}{string of folders containing table - separate folders with '/'}

\item{name}{the name of the table (csv extension is not required)}

\item{type}{the table type i.e. results (default) or analysis etc.}

\item{digits}{count of number of significant digits}
}
\value{
a data.frame of the table
}
\description{
Reads in csv file from current table folder.
}

