\docType{package}
\name{package-poiscon}
\alias{package-poiscon}
\alias{package-poiscon-package}
\alias{poiscon}
\title{Poiscon}
\description{
Utility functions to simplify date time manipulations and
data, plot, table and analysis saving and loading. It is
used by the employees of Poisson Consulting Ltd.
}

