\name{parent_folder}
\alias{parent_folder}
\title{Parent folder}
\usage{
parent_folder()
}
\value{
The name of the parent folder as a character scalar.
}
\description{
Returns the name of the parent folder. It's uses include
determining the name of the current project (the parent
directory should have the same name as the associated
github repository).
}

