\name{plot_bch_data}
\alias{plot_bch_data}
\title{Plotting BCH data}
\usage{
  plot_bch_data(data, file = NULL)
}
\arguments{
  \item{data}{a data.frame of the BCH data to plot.}

  \item{file}{a character scalar or NULL of the file name
  for the plot.}
}
\value{
  The data is written to a pdf file.
}
\description{
  Plots environmental data
}

