% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/perform-analyses.R
\name{plot_residuals_analyses}
\alias{plot_residuals_analyses}
\title{Plots Residuals}
\usage{
plot_residuals_analyses(...)
}
\arguments{
\item{...}{character vectors of folders to cycle through}
}
\description{
Plots
}

