% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-plot.R
\name{print_plot}
\alias{print_plot}
\title{Print Plot}
\usage{
print_plot(x)
}
\arguments{
\item{x}{The plot object to plot.}
}
\description{
Function to use an print_plot(gp) in place of print(gp) as catches error and retries.
}

