\name{project_folder}
\alias{project_folder}
\title{Project folder}
\usage{
project_folder()
}
\value{
The name of the project folder as a character scalar.
}
\description{
Returns the name of the parent folder. It's uses include
determining the name of the current project (the parent
directory should have the same name as the associated
github repository).
}

