\name{read_cpc_data}
\alias{read_cpc_data}
\title{Read CPC data}
\usage{
read_cpc_data(file, recursive = TRUE, quiet = TRUE)
}
\arguments{
  \item{file}{the location of the zxrp file or folder with
  zxrp files}

  \item{recursive}{a logical scalar indicating whether to
  read data from subdirectories}

  \item{quiet}{a logical scalar indicating whether to
  provide messages}
}
\value{
the data in the form of a data.frame
}
\description{
Reads CPC data into R
}

