\name{read_square_data}
\alias{read_square_data}
\title{read_square_data}
\usage{
  read_square_data(file = "data.csv")
}
\arguments{
  \item{file}{the file with the square data is comma
  separated format}
}
\value{
  the data in the form of a object of class hdata
}
\description{
  Read square data into R
}

