\name{read_zrxp}
\alias{read_zrxp}
\title{Read zrxp data into R}
\usage{
  read_zrxp(file, recursive = FALSE, quiet = TRUE)
}
\arguments{
  \item{file}{the location of the zxrp file or folder with
  zxrp files}
}
\value{
  the data in the form of a data.frame
}
\description{
  Reads zrxp data into R
}

