\name{read_zrxp_data}
\alias{read_zrxp_data}
\title{read_zrxp_data}
\usage{
  read_zrxp_data(file = "data.zxrp",
    value.name = "Discharge", location = "")
}
\arguments{
  \item{file}{the zxrp file with the data}

  \item{value.name}{the name of the variable}

  \item{location}{the name of the location}
}
\value{
  the data in the form of a data.frame
}
\description{
  Read zrxp data into R
}

