% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util.R
\name{remove_dots_colnames_data_frames}
\alias{remove_dots_colnames_data_frames}
\title{Remove Dots Colnames}
\usage{
remove_dots_colnames_data_frames()
}
\description{
Goes through all the data.frame objects in
the current environment and removes any dots
from the colnames
}

