\name{report_url}
\alias{report_url}
\title{Poisson report URL}
\usage{
report_url(date)
}
\arguments{
  \item{date}{Date scalar of when the report was posted.}
}
\value{
The url of the poisson report.
}
\description{
Returns the URL of the poisson report corresponding to the
current project and date of posting.
}

