\name{resave_plots}
\alias{resave_plots}
\title{Resave plots}
\usage{
resave_plots(type = "figures")
}
\arguments{
  \item{type}{a character element indicating the type of
  plots to redraw by default type = "figures"}
}
\value{
invisible vector of files resaved
}
\description{
Can be used to resave plots with different theme and/or if
in windows to ensure correct height and width when inserted
into Microsoft Word document#'
}

