\name{reset_dirs}
\alias{reset_dirs}
\title{Reset poiscon directories}
\usage{
  reset_dirs(share = FALSE)
}
\arguments{
  \item{share}{a logical element indicating whether or not
  the directories should be nested in the work directory
  (share = TRUE) or saved a level higher in the directory
  structure (share = FALSE) the default}
}
\value{
  NULL
}
\description{
  Resets poiscon directories to the default values. The
  directories are used for saving plots, tables, rdata and
  analyses.
}

