% Generated by roxygen2 (4.0.1): do not edit by hand
\name{reset_folders}
\alias{reset_folders}
\title{Reset poiscon folders}
\usage{
reset_folders(dir = "output")
}
\arguments{
\item{dir}{a character scalar indicating the name of the folders}
}
\description{
Resets poiscon folders to the default values. The folders are
used for saving plots, tables, rdata and analyses.
}

