\name{s}
\alias{s}
\title{Summary}
\usage{
  s(object, ...)
}
\arguments{
  \item{object}{an object to summarise}

  \item{...}{further arguments passed to or from other
  methods.}
}
\value{
  summarises x.
}
\description{
  Short hand for summary
}
\seealso{
  \code{\link{p}}, \code{\link{h}}, \code{\link{ht}}.
}

