\name{save_plots.data.frame}
\alias{save_plots.data.frame}
\title{Save plots}
\usage{
  \method{save_plots}{data.frame} (object, name = "data",
    ...)
}
\arguments{
  \item{object}{data.frame object to plot}

  \item{name}{a character scalar of the file name.}

  \item{...}{further arguments passed to or from other
  methods.}
}
\value{
  Saves data.frame as a pdf file.
}
\description{
  Plots data.frame object to a pdf file
}

