% Generated by roxygen2 (4.0.2): do not edit by hand
\name{save_plots.jags_analysis}
\alias{save_plots.jags_analysis}
\title{Save plots}
\usage{
\method{save_plots}{jags_analysis}(object, model = 1, ...)
}
\arguments{
\item{object}{jags_analysis object to plot}

\item{model}{a number or string of the model to plot}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Saves jags_analysis as a pdf file.
}
\description{
Plots jags_analysis object to a pdf file
}

