% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save-plots.r
\name{save_plots}
\alias{save_plots}
\title{Save plots}
\usage{
save_plots(object, ...)
}
\arguments{
\item{object}{object to plot}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Save plot as a pdf file.
}
\description{
Plots object to a pdf file
}
