\name{save_rdata}
\alias{save_rdata}
\title{Save rdata}
\usage{
save_rdata(object, name = "data")
}
\arguments{
  \item{object}{the object to save.}

  \item{name}{a character scalar of the name of the file.}
}
\value{
Saves object as .rdata files in current rdata folder.
}
\description{
Saves object as .rdata files in current rdata folder.
}

