\name{save_table}
\alias{save_table}
\title{Save table}
\usage{
save_table(object, name = "table", type = "results", row.names = FALSE)
}
\arguments{
  \item{object}{the object to save.}

  \item{name}{a character scalar of the name of the
  object.}

  \item{type}{a character scalar of the type of folder to
  save in.}

  \item{row.names}{a logical scalar of whether to include
  row names.}
}
\value{
Saves object as .csv files in current table folder.
}
\description{
Saves object as .csv file in current table folder.
}

