% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save-tables.r
\name{save_tables.jags_analysis}
\alias{save_tables.jags_analysis}
\title{Save tables}
\usage{
\method{save_tables}{jags_analysis}(object, model = 1, ...)
}
\arguments{
\item{object}{jags_analysis object to tables}

\item{model}{a integer scalar of the model number}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Save object as a tables.
}
\description{
Saves jags_analysis to tables
}

