% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save-tables.r
\name{save_tables}
\alias{save_tables}
\title{Save tables}
\usage{
save_tables(object, ...)
}
\arguments{
\item{object}{object to save tables}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Save object as a tables.
}
\description{
Saves object to a tables file
}
