% Generated by roxygen2 (4.0.1): do not edit by hand
\name{set_colors}
\alias{set_colors}
\title{Sets colors}
\usage{
set_colors(values = get_colors())
}
\arguments{
\item{values}{character vector of colors}
}
\value{
invisible character vector of colors
}
\description{
Sets color palette for poiscon
}

