% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-folders.r
\name{set_folders}
\alias{set_folders}
\title{Set folders}
\usage{
set_folders(...)
}
\arguments{
\item{...}{multiple character scalars}
}
\description{
Sets folders for rdata, plots, tables, and analyses
}

