\name{theme_Golder}
\alias{theme_Golder}
\title{Theme Golder}
\usage{
theme_Golder(base_size = 10, base_family = "Arial")
}
\arguments{
  \item{base_size}{an integer scalar indicating the base
  font size.}

  \item{base_family}{a character scalar indicating the base
  font family.}
}
\value{
A ggplot2 theme.
}
\description{
A ggplot2 theme for Golder plots
}

