\name{upload_web}
\alias{upload_web}
\title{Upload web}
\usage{
upload_web(dir = "poissonconsulting.github.io", path = options()$code_dir,
  message = paste0(project_folder(), " ", Sys.time()))
}
\arguments{
  \item{web_dir}{a character scalar or a Date}
}
\value{
Uploads files to jekyll site on my harddrive.
}
\description{
Uploads project report files to jekyll site on my
harddrive.
}

