\name{working_dir}
\alias{working_dir}
\title{Working directory folder}
\usage{
working_dir()
}
\value{
The name of the project folder as a character scalar.
}
\description{
Returns the name of the parent folder. It's uses include
determining the name of the current project (the parent
directory should have the same name as the associated
github repository).
}

