\name{write_bch_data}
\alias{write_bch_data}
\title{Write BCH data}
\usage{
write_bch_data(data, file = NULL)
}
\arguments{
  \item{data}{the BCH data to write}

  \item{file}{NULL or a character scalar of the files name
  without ".csv"}
}
\value{
The results of write.csv
}
\description{
Writes BCH data to csv files - creating file name from the
date if unspecified
}

