% Generated by roxygen2 (4.0.2): do not edit by hand
\name{delete_output}
\alias{delete_output}
\title{Delete output folder}
\usage{
delete_output(delete_rdata_input = FALSE, check = TRUE)
}
\arguments{
\item{delete_rdata_input}{flag indicating whether to delete the output/rdata/input
folder which in the PCL system is used to store the initial input of data}

\item{check}{flag indicating whether to confirm with user before deleting}
}
\value{
An invisible logical scalar indicating if successful.
}
\description{
Deletes output folder for current project.
}

