\name{gwindow}
\alias{gwindow}
\title{Plotting window}
\usage{
gwindow(width = 100, height = NULL)
}
\arguments{
  \item{width}{a numeric scalar indicating the percent of
  the page width (or the number of plots per page width -
  up to 10)}

  \item{height}{a numeric scalar indicating the percent of
  the page height - by default the same percent as the
  width (or the number of plots per page width - up to
  10).}
}
\value{
Creates a new window for plotting
}
\description{
Creates a new platform-independent window for plotting
}

