\name{gwindow}
\alias{gwindow}
\title{Plotting window}
\usage{
  gwindow(dwidth = 1, dheight = 1, slide = FALSE)
}
\arguments{
  \item{dwidth}{a numeric scalar indicating the proportion
  of the page width}

  \item{dheight}{a numeric scalar indicating the proprotion
  of the page height}

  \item{slide}{a logical scalar indicating whether plot
  (the default) versus powerpoint slide.}
}
\value{
  Creates a new window for plotting
}
\description{
  Creates a new platform-independent window for plotting
}

