% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwindow.r
\name{gwindow}
\alias{gwindow}
\title{Open graphics window}
\usage{
gwindow(width = 100, height = width)
}
\arguments{
\item{width}{a numeric scalar indicating the percent of the page width (or
if 10 or less the width of the page in inches)}

\item{height}{a numeric scalar indicating the percent of the page width (or
if 10 or less the width of the page in inches). By default height is
the same as the width.}
}
\value{
a new graphics window for plotting
}
\description{
Opens a new graphics window for plotting by default the page width is
6 inches.
}
\examples{
\dontrun{

gwindow(50)
gwindow(3)
gwindow(50, 100)
gwindow(3, 6)
gwindow(3, 100)
}
}
\seealso{
\code{\link{save_plot}}
}
