\name{is_predictive}
\alias{is_predictive}
\title{Is Predictive}
\usage{
is_predictive(x)
}
\arguments{
  \item{x}{data.frame of predictions to test informative}
}
\value{
logical scalar indicating whether informative
}
\description{
Tests whether a prediction data.frame is informative i.e.,
has variation in estimate and/or lower/upper.
}

