% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/load-table.R
\name{load_table}
\alias{load_table}
\title{Load table}
\usage{
load_table(name = "table", type = "results")
}
\arguments{
\item{name}{the name of the table (csv extension is not required)}

\item{type}{the table type i.e. results (default) or analysis etc.}
}
\value{
a data.frame of the table
}
\description{
Reads in csv file from current table folder.
}

