% Generated by roxygen2 (4.0.2): do not edit by hand
\name{load_value}
\alias{load_value}
\title{Load table}
\usage{
load_value(colname, rowname, folders = get_folders(), name = "table",
  type = "results")
}
\arguments{
\item{colname}{string of name of column containing value}

\item{rowname}{string of name of row containing value}

\item{folders}{string of folders containing table - separate folders with '/'}

\item{name}{the name of the table (csv extension is not required)}

\item{type}{the table type i.e. results (default) or analysis etc.}
}
\value{
a data.frame of the table
}
\description{
Reads in csv file from current table folder.
}

