% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{package-poiscon}
\alias{package-poiscon}
\alias{package-poiscon-package}
\alias{poiscon}
\title{Poiscon}
\description{
Utility functions to simplify date time manipulations and
data, plot, table and analysis saving and loading. It is used
by the employees of Poisson Consulting Ltd.
}

