% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.r
\docType{package}
\name{package-poiscon}
\alias{package-poiscon}
\alias{poiscon}
\alias{package-poiscon-package}
\title{Poiscon}
\description{
Utility functions to simplify date time manipulations and
data, plot, table and analysis saving and loading. It is used
by the employees of Poisson Consulting Ltd.
}
