% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/perform-analyses.R
\name{perform_analyses}
\alias{perform_analyses}
\title{Perform Analyses}
\usage{
perform_analyses(models, ..., niters = 10^3, mode = "current",
  name = "data", beep = TRUE)
}
\arguments{
\item{models}{jags_model objects}

\item{...}{character vectors of folders to cycle through}

\item{niters}{count of number of iterations}

\item{mode}{string of opts_jagr model}

\item{name}{string of name of data file}

\item{beep}{flag indicating whether to beep on completion}
}
\description{
Performs batch jags analyses
}

