% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform-analyses.R
\name{perform_analyses}
\alias{perform_analyses}
\title{Perform Analyses}
\usage{
perform_analyses(models, ..., niters = 10^3, mode = "current",
  name = "data", beep = getOption("poiscon.beep", default = FALSE))
}
\arguments{
\item{models}{A jags_model object.}

\item{...}{A series of character vectors or a list of character vectors of 
folders to cycle through.}

\item{niters}{A count of the number of iterations to perform.}

\item{mode}{A string of the opts_jagr mode.}

\item{name}{A string of the name of the data file.}

\item{beep}{A flag indicating whether to beep on completion.}
}
\description{
Performs batch jags analyses. It goes through the folders defined by \code{...}
and pulls out the data of name \code{name} (by default \code{"data"} and then fits
the models defined in \code{models}. Note \code{beep = TRUE} has caused
crashed with RStudio on windows.
}

