\name{read_zrxp_data}
\alias{read_zrxp}
\alias{read_zrxp_data}
\title{Read zrxp data into R}
\usage{
read_zrxp_data(file, recursive = FALSE, quiet = TRUE)
}
\arguments{
  \item{file}{the location of the zxrp file or folder with
  zxrp files}

  \item{recursive}{a logical scalar indicating whether to
  read data from subdirectories}

  \item{quiet}{a logical scalar indicating whether to
  provide messages}
}
\value{
the data in the form of a data.frame
}
\description{
Reads zrxp data into R
}

