% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/save.r
\name{save_rdata}
\alias{save_rdata}
\title{Save rdata}
\usage{
save_rdata(object, name = NULL)
}
\arguments{
\item{object}{the object to save if unspecified the
function saves all data.frames in the global environment}

\item{name}{string of the name of the file if unspecified
the file is named the same as the object. If object is unspecified
name is ignored and each data.frame in the global environment is saved.}
}
\value{
Invisible flag indicating if successful (also issues warning if FALSE).
}
\description{
Saves object as .rdata files in current rdata folder.
}

