% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/save-table.R
\name{save_table}
\alias{save_table}
\title{Save table}
\usage{
save_table(x, name = "table", type = "results")
}
\arguments{
\item{x}{the data.frame to to save}

\item{name}{character scalar of the name for the csv table}

\item{type}{character scalar of the type of folder to save in}
}
\value{
Saves x as csv file in current table folder.
}
\description{
Saves data.frame as csv file in current table folder. Row names
are only save if informative (i.e. not NULL and not just 1:nrow(x)).
}

