\name{set_colors}
\alias{set_colors}
\title{Sets colors}
\usage{
set_colors(values = c("#000000", "#E69F00", "#56B4E9", "#009E73", "#F0E442",
  "#0072B2", "#D55E00", "#CC79A7"))
}
\arguments{
  \item{values}{character vector of colors (by default a
  color blind friendly palette starting with black)}
}
\value{
invisible logical scalar indicating if successful
}
\description{
Sets color palette for poiscon
}

