% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardise-datetime.r
\name{standardised_date_time}
\alias{standardised_date_time}
\title{Standardised date time}
\usage{
standardised_date_time(x, standardised_offset = -8)
}
\arguments{
\item{x}{a date-time object}

\item{standardised_offset}{an integer scalar indicating the time difference
in hours between UTC and the desired time zone.}
}
\description{
Deprecated. See  \code{\link{standardise_datetime}}
}
