\name{theme_Poisson}
\alias{theme_Poisson}
\title{Theme Poisson}
\usage{
theme_Poisson(base_size = 10, base_family = "Arial")
}
\arguments{
  \item{base_size}{an integer scalar indicating the base
  font size.}

  \item{base_family}{a character scalar indicating the base
  font family.}
}
\value{
A ggplot2 theme.
}
\description{
A ggplot2 theme for Poisson plots
}

