% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-Poisson.r
\name{theme_Poisson}
\alias{theme_Poisson}
\title{Theme Poisson}
\usage{
theme_Poisson(base_size = 12, base_family = "")
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}
}
\value{
A gg theme object.
}
\description{
A gg theme for Poisson style plots.
}
\details{
To use, ensure ggplot2 is installed and then call
\code{theme_set(theme_Poisson())}.
}
\examples{
library(ggplot2)
ggplot(data = mpg, aes(x = cty, y = hwy)) +
facet_wrap(~manufacturer) +
geom_count() +
theme_Poisson()
}

